+++ sdcard/rootfs/root/pwnagotchi/scripts/pwnagotchi/agent.py
@@ -9,6 +9,7 @@ import _thread
 
 import core
 
+import pwnagotchi.utils as utils
 import pwnagotchi.plugins as plugins
 from bettercap.client import Client
 from pwnagotchi.mesh.utils import AsyncAdvertiser
@@ -30,7 +31,7 @@ class Agent(Client, AsyncAdvertiser, AsyncTrainer):
         self._started_at = time.time()
         self._filter = None if config['main']['filter'] is None else re.compile(config['main']['filter'])
         self._current_channel = 0
-        self._supported_channels = core.iface_channels(config['main']['iface'])
+        self._supported_channels = utils.iface_channels(config['main']['iface'])
         self._view = view
         self._access_points = []
         self._last_pwnd = None
@@ -252,7 +253,7 @@ class Agent(Client, AsyncAdvertiser, AsyncTrainer):
 
     def _update_uptime(self, s):
         secs = time.time() - self._started_at
-        self._view.set('uptime', core.secs_to_hhmmss(secs))
+        self._view.set('uptime', utils.secs_to_hhmmss(secs))
         self._view.set('epoch', '%04d' % self._epoch.epoch)
 
     def _update_counters(self):
@@ -272,7 +273,7 @@ class Agent(Client, AsyncAdvertiser, AsyncTrainer):
         if new_shakes > 0:
             self._epoch.track(handshake=True, inc=new_shakes)
 
-        tot = core.total_unique_handshakes(self._config['bettercap']['handshakes'])
+        tot = utils.total_unique_handshakes(self._config['bettercap']['handshakes'])
         txt = '%d (%d)' % (len(self._handshakes), tot)
 
         if self._last_pwnd is not None:
@@ -285,7 +286,7 @@ class Agent(Client, AsyncAdvertiser, AsyncTrainer):
 
     def _update_advertisement(self, s):
         run_handshakes = len(self._handshakes)
-        tot_handshakes = core.total_unique_handshakes(self._config['bettercap']['handshakes'])
+        tot_handshakes = utils.total_unique_handshakes(self._config['bettercap']['handshakes'])
         started = s['started_at'].split('.')[0]
         started = datetime.strptime(started, '%Y-%m-%dT%H:%M:%S')
         started = time.mktime(started.timetuple())
