+++ sdcard/rootfs/root/pwnagotchi/scripts/pwnagotchi/ai/epoch.py
@@ -2,8 +2,8 @@ import time
 import threading
 import logging
 
-import core
 import pwnagotchi
+import pwnagotchi.utils as utils
 import pwnagotchi.mesh.wifi as wifi
 
 from pwnagotchi.ai.reward import RewardFunction
@@ -174,22 +174,22 @@ class Epoch(object):
         self._epoch_data_ready.set()
 
         logging.info("[epoch %d] duration=%s slept_for=%s blind=%d inactive=%d active=%d hops=%d missed=%d "
-                 "deauths=%d assocs=%d handshakes=%d cpu=%d%% mem=%d%% temperature=%dC reward=%s" % (
-                     self.epoch,
-                     core.secs_to_hhmmss(self.epoch_duration),
-                     core.secs_to_hhmmss(self.num_slept),
-                     self.blind_for,
-                     self.inactive_for,
-                     self.active_for,
-                     self.num_hops,
-                     self.num_missed,
-                     self.num_deauths,
-                     self.num_assocs,
-                     self.num_shakes,
-                     cpu * 100,
-                     mem * 100,
-                     temp,
-                     self._epoch_data['reward']))
+                     "deauths=%d assocs=%d handshakes=%d cpu=%d%% mem=%d%% temperature=%dC reward=%s" % (
+                         self.epoch,
+                         utils.secs_to_hhmmss(self.epoch_duration),
+                         utils.secs_to_hhmmss(self.num_slept),
+                         self.blind_for,
+                         self.inactive_for,
+                         self.active_for,
+                         self.num_hops,
+                         self.num_missed,
+                         self.num_deauths,
+                         self.num_assocs,
+                         self.num_shakes,
+                         cpu * 100,
+                         mem * 100,
+                         temp,
+                         self._epoch_data['reward']))
 
         self.epoch += 1
         self.epoch_started = now
