+++ sdcard/rootfs/root/pwnagotchi/scripts/pwnagotchi/ui/view.py
@@ -4,7 +4,7 @@ import time
 import logging
 from PIL import Image, ImageDraw
 
-import core
+import pwnagotchi.utils as utils
 import pwnagotchi.plugins as plugins
 from pwnagotchi.voice import Voice
 
@@ -86,8 +86,9 @@ class View(object):
 
             'face': Text(value=faces.SLEEP, position=face_pos, color=BLACK, font=fonts.Huge),
 
-            'friend_face': Text(value=None, position=(0, (self._height * 0.88) - 15 ), font=fonts.Bold, color=BLACK),
-            'friend_name': Text(value=None, position=(40,(self._height * 0.88) - 13 ), font=fonts.BoldSmall, color=BLACK),
+            'friend_face': Text(value=None, position=(0, (self._height * 0.88) - 15), font=fonts.Bold, color=BLACK),
+            'friend_name': Text(value=None, position=(40, (self._height * 0.88) - 13), font=fonts.BoldSmall,
+                                color=BLACK),
 
             'name': Text(value='%s>' % 'pwnagotchi', position=name_pos, color=BLACK, font=fonts.Bold),
 
@@ -166,7 +167,7 @@ class View(object):
         self.set('channel', '-')
         self.set('aps', "%d" % log.associated)
         self.set('shakes', '%d (%s)' % (log.handshakes, \
-                                        core.total_unique_handshakes(self._config['bettercap']['handshakes'])))
+                                        utils.total_unique_handshakes(self._config['bettercap']['handshakes'])))
         self.set_closest_peer(log.last_peer)
 
     def is_normal(self):
