+++ sdcard/rootfs/root/pwnagotchi/scripts/pwnagotchi/utils.py
@@ -1,6 +1,9 @@
-import yaml
-import os
 import logging
+import glob
+import os
+import time
+import subprocess
+import yaml
 
 
 # https://stackoverflow.com/questions/823196/yaml-merge-in-python
@@ -40,3 +43,38 @@ def setup_logging(args, config):
     console_handler = logging.StreamHandler()
     console_handler.setFormatter(formatter)
     root.addHandler(console_handler)
+
+
+def secs_to_hhmmss(secs):
+    mins, secs = divmod(secs, 60)
+    hours, mins = divmod(mins, 60)
+    return '%02d:%02d:%02d' % (hours, mins, secs)
+
+
+def total_unique_handshakes(path):
+    expr = os.path.join(path, "*.pcap")
+    return len(glob.glob(expr))
+
+
+def iface_channels(ifname):
+    channels = []
+    output = subprocess.getoutput("/sbin/iwlist %s freq" % ifname)
+    for line in output.split("\n"):
+        line = line.strip()
+        if line.startswith("Channel "):
+            channels.append(int(line.split()[1]))
+    return channels
+
+
+def led(on=True):
+    with open('/sys/class/leds/led0/brightness', 'w+t') as fp:
+        fp.write("%d" % (0 if on is True else 1))
+
+
+def blink(times=1, delay=0.3):
+    for t in range(0, times):
+        led(True)
+        time.sleep(delay)
+        led(False)
+        time.sleep(delay)
+    led(True)
